requirements.txt
numpy
torch
Pillow
opencv-python
torchvision